+++ httpcore/_backends/anyio.py
@@ -10,6 +10,7 @@ from .._exceptions import (
     ConnectTimeout,
     ReadError,
     ReadTimeout,
+    SSLError,
     WriteError,
     WriteTimeout,
     map_exceptions,
@@ -62,7 +63,7 @@ class AnyIOStream(AsyncNetworkStream):
             TimeoutError: ConnectTimeout,
             anyio.BrokenResourceError: ConnectError,
             anyio.EndOfStream: ConnectError,
-            ssl.SSLError: ConnectError,
+            ssl.SSLError: SSLError,
         }
         with map_exceptions(exc_map):
             try: